refactor: group entities into entities folder
