+++ server/src/entities/Base.entity.ts
