+++ server/src/entities/User.entity.ts
@@ -1,5 +1,5 @@
 import { Exclude } from 'class-transformer';
-import { BaseEntity } from 'src/shared/Base/base.entity';
+import { BaseEntity } from './Base.entity';
 import { Column, Entity } from 'typeorm';
 
 @Entity()
