+++ server/src/modules/message/messages.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { Message } from './messages.entity';
+import { Message } from 'src/entities/Message.entity';
 import { MessagesController } from './messages.controller';
 import { MessagesService } from './messages.service';
