+++ server/src/modules/message/messages.service.ts
@@ -2,7 +2,7 @@ import { Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 
-import { Message } from './messages.entity';
+import { Message } from 'src/entities/Message.entity';
 
 @Injectable()
 export class MessagesService {
