+++ server/src/shared/Decorators/user-info.decorator.ts
@@ -1,5 +1,5 @@
 import { createParamDecorator, ExecutionContext } from '@nestjs/common';
-import { User } from 'src/modules/users/users.entity';
+import { User } from 'src/entities/User.entity';
 
 export const UserInfo = createParamDecorator(
   (data, ctx: ExecutionContext): User => {
